requirements.txt
asgiref==3.5.2
autopep8==1.7.0
backports.zoneinfo==0.2.1
Django==4.1.2
django-searchview-lib==1.0.5
Pillow==9.2.0
pycodestyle==2.9.1
sqlparse==0.4.3
toml==0.10.2
